# server/app/static/openapi.yaml
openapi: 3.1.0
info:
  title: QueryBot API
  version: 1.0.0
  description: |
    Natural-language analytics over uploaded datasets.

    ## Authentication
    Sign-in sets two httpOnly cookies, `access_token` and `refresh_token`. The
    browser attaches them automatically; no token is ever readable from
    JavaScript. When the access token expires the client calls
    `POST /auth/refresh` once and replays the original request.

    ## Authorization
    Datasets belong to the user who uploaded them. Every endpoint that reads a
    dataset verifies ownership and answers `404` — not `403` — for a dataset
    owned by someone else, so identifiers cannot be probed for existence.

    ## Rate limits
    `POST /langgraph/run` calls a paid model several times per question and is
    limited to 20 per hour per user. Responses carry `X-RateLimit-*` headers.
  license:
    name: ISC
servers:
  - url: http://localhost:8080/api
    description: Docker Compose
  - url: http://localhost:5000/api
    description: Local development
...
components:
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: access_token
  schemas:
    Success:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data: { type: object }
    Error:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean, const: false }
        message: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
...
    Dataset:
      type: object
      properties:
        id: { type: string, format: uuid }
        uuid:
          type: string
          format: uuid
          description: Identifier used when querying this dataset.
        file_name: { type: string }
        size_bytes: { type: integer }
        table_count: { type: integer }
        row_count: { type: integer }
        expires_at:
          type: [string, 'null']
          format: date-time
          description: Uploads are deleted after the retention window.
    Table:
      type: object
      properties:
        name: { type: string }
        rowCount: { type: integer }
        columns:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              type: { type: string }
              nullable: { type: boolean }
              primaryKey: { type: boolean }
    ColumnProfile:
      type: object
      properties:
        name: { type: string }
        type: { type: string, description: The type SQLite declares. }
        kind:
          type: string
          enum: [numeric, text, temporal, boolean]
          description: Inferred from the values, since a CSV date column arrives as TEXT.
        nullCount: { type: integer }
        nullPercent: { type: number }
        distinctCount: { type: integer }
        isUnique: { type: boolean }
        isIdentifier:
          type: boolean
          description: >
            Identifies rows rather than measuring anything - a declared primary
            key, or a unique integer with an identifier-like name. Excluded from
            correlations, where it would only produce artefacts of insertion order.
        min: { type: number }
        max: { type: number }
        mean: { type: number }
        median: { type: number }
        stdev: { type: number }
        outlierCount:
          type: integer
          description: Values beyond three interquartile ranges from the middle 50%.
        topValues:
          type: array
          description: Present for low-cardinality non-numeric columns.
          items:
            type: object
            properties:
              value: { type: string }
              count: { type: integer }
        histogram:
          type: array
          description: Bucketed counts for numeric columns.
          items:
            type: object
            properties:
              start: { type: number }
              end: { type: number }
              count: { type: integer }
...
  responses:
    Unauthorized:
      description: No valid session.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found, or owned by another user.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Too many requests.
      headers:
        X-RateLimit-Remaining: { schema: { type: integer } }
        Retry-After: { schema: { type: integer } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
security:
  - accessCookie: []
paths:
...
  /datasets:
    get:
      tags: [Datasets]
      summary: List the caller's datasets
      responses:
        '200':
          description: Datasets, most recently used first.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          datasets:
                            type: array
                            items: { $ref: '#/components/schemas/Dataset' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Datasets]
      summary: Upload a dataset
      description: |
        Accepts `.csv`, `.sqlite` and `.db`, up to 100MB. CSV files are converted
        to SQLite with column types inferred from the data, so numeric columns
        sort and aggregate correctly rather than comparing as text.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '201':
          description: Uploaded.
        '400': { description: Unsupported or unreadable file. }
        '413': { description: File too large. }
        '429': { $ref: '#/components/responses/RateLimited' }
  /datasets/{uuid}/schema:
    get:
      tags: [Datasets]
      summary: Get tables and columns
      parameters:
        - name: uuid
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Structured schema.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          tables:
                            type: array
                            items: { $ref: '#/components/schemas/Table' }
        '404': { $ref: '#/components/responses/NotFound' }
  /datasets/{uuid}/profile:
    get:
      tags: [Datasets]
      summary: Statistically profile a dataset
      description: |
        Per-column statistics computed by scanning the data: type inference,
        missing-value counts, cardinality, min/max/mean/median/stdev, outlier
        counts, value distributions, duplicate rows, and Pearson correlations
        between numeric columns.

        Computed entirely in SQL — no model call — so it is fast and exact, and
        cached for the dataset's lifetime since the underlying data is immutable
        once uploaded.

        `highlights` contains plain-language observations worth acting on, such as
        a column that is mostly empty, one that holds a single value throughout,
        or a distribution skewed enough that the median is the fairer summary.
      parameters:
        - name: uuid
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: The dataset profile.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          tables:
                            type: array
                            items:
                              type: object
                              properties:
                                table: { type: string }
                                rowCount: { type: integer }
                                duplicateRows: { type: integer }
                                columns:
                                  type: array
                                  items: { $ref: '#/components/schemas/ColumnProfile' }
                          correlations:
                            type: array
                            items:
                              type: object
                              properties:
                                a: { type: string }
                                b: { type: string }
                                coefficient: { type: number }
                          highlights:
                            type: array
                            items: { type: string }
                          computedMs: { type: integer }
        '404': { $ref: '#/components/responses/NotFound' }
  /datasets/{uuid}/preview/{table}:
    get:
      tags: [Datasets]
      summary: Preview the first rows of a table
      parameters:
        - name: uuid
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: table
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Preview rows. }
        '404': { $ref: '#/components/responses/NotFound' }
  /datasets/{uuid}:
    delete:
      tags: [Datasets]
      summary: Delete a dataset
      parameters:
        - name: uuid
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Deleted. }
        '404': { $ref: '#/components/responses/NotFound' }
...